// src/utils/enums.ts
export enum EventType {
  RECOMMENDED = "recommended",
  PAST = "past",
  UPCOMMING = "upcomming",
  PREMIUM = "premium",
  DEFAULT = "default",
  TRENDING = "trending",
  POPULAR = "popular",
  ALL = "all",
  SEARCH = "search",
}

export enum EventFlow {
  DEFAULT = "default",
  STANDING = "standing",
  SITTING = "sitting",
}

export enum STATUS {
  DEFAULT = "default",
  ACTIVE = "active",
  IN_ACTIVATED = "in-active",
  PENDING = "pending",
  DELETED = "delete",
  COMPLETED = "complete",
  POSTPONED = "postponed",
  CANCELLED = "cancelled",
}

export enum DiscountType {
  PERCENTAGE = "percentage",
  FIXED_AMOUNT = "fixed_amount",
  DEFAULT = "default",
}
export enum CouponType {
  EVENTS = "events",
  USERS = "users",
  GLOBAL = "global",
}